+++ internal/conformance/cmd/test_booleans/main.go
@@ -9,7 +9,7 @@ import (
 
 func main() {
 	n := conformance.GetTestCases()
-	hegel.MustRun(func(s *hegel.TestCase) {
+	hegel.MustRun(func(s hegel.TestCase) {
 		defer conformance.EnsureMetric()
 		v := hegel.Draw(s, hegel.Booleans())
 		conformance.WriteMetrics(map[string]any{