+++ internal/conformance/cmd/test_sampled_from/main.go
@@ -24,7 +24,7 @@ func main() {
 
 	gen := hegel.SampledFrom(params.Options)
 	n := conformance.GetTestCases()
-	hegel.MustRun(func(s *hegel.TestCase) {
+	hegel.MustRun(func(s hegel.TestCase) {
 		defer conformance.EnsureMetric()
 		val := hegel.Draw(s, gen)
 		conformance.WriteMetrics(map[string]any{